requirements/base.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
cookies==2.2.1
defusedxml==0.4.1
Django==1.9.7
django-allauth==0.25.2
django-celery==3.1.17
django-tastypie==0.13.3
django-timezone-field==1.3
geonames==0.1.3
kombu==3.0.35
oauthlib==1.1.2
psycopg2==2.6.2
python-dateutil==2.5.3
python-forecastio==1.3.4
python-mimeparse==1.5.2
python3-openid==3.0.10
pytz==2016.6.1
requests==2.10.0
requests-oauthlib==0.6.2
responses==0.5.1
six==1.10.0
tzwhere==2.3
unicodecsv==0.14.1
